Move dashboard routes to own router
