+++ src/server/trpc/router/profile.ts
@@ -1,12 +1,6 @@
 import { z } from "zod";
-import getDateXDaysAgo from "../../../utils/getDateXAgo";
 import paginate from "../../../utils/paginate";
 import { router, protectedProcedure, publicProcedure } from "../trpc";
-
-interface IStatItem {
-  date: string;
-  count: number;
-}
 
 export const profileRouter = router({
   profileBySession: protectedProcedure.query(async ({ ctx }) => {
@@ -169,84 +163,6 @@ export const profileRouter = router({
 
       return user;
     }),
-
-  stats: protectedProcedure.query(async ({ ctx }) => {
-    const items: {
-      date: string;
-      count: number;
-    }[] = [];
-    let episodeCounter = 0;
-    let movieCounter = 0;
-
-    for (let i = 0; i < 14; i++) {
-      const gte = new Date(getDateXDaysAgo(i).setHours(0, 0, 0, 0));
-      const lt = new Date(getDateXDaysAgo(i - 1).setHours(0, 0, 0, 0));
-
-      await ctx.prisma.episodesHistory
-        .count({
-          where: {
-            user_id: ctx.session.user.profile.userId,
-            datetime: {
-              gte,
-              lt,
-            },
-          },
-        })
-        .then((res) => {
-          if (res > 0) episodeCounter += res;
-          items.push({
-            date: gte.toLocaleString("en-UK", {
-              month: "short",
-              day: "numeric",
-            }),
-            count: res,
-          });
-        });
-
-      await ctx.prisma.moviesHistory
-        .count({
-          where: {
-            user_id: ctx.session.user.profile.userId,
-            datetime: {
-              gte,
-              lt,
-            },
-          },
-        })
-        .then((res) => {
-          if (res > 0) movieCounter++;
-          items.push({
-            date: gte.toLocaleString("en-UK", {
-              month: "short",
-              day: "numeric",
-            }),
-            count: res,
-          });
-        });
-    }
-
-    const result: IStatItem[] = items.reduce((acc: any, { date, count }: any) => {
-      acc[date] ??= { date: date, count: [] };
-      acc[date].count = Number(acc[date].count) + Number(count);
-
-      return acc;
-    }, {});
-
-    const sorted = Object.values(result).sort((a: IStatItem, b: IStatItem) => {
-      if (new Date(a.date) > new Date(b.date)) {
-        return 1;
-      } else {
-        return -1;
-      }
-    });
-
-    return {
-      history: sorted,
-      episodeAmount: episodeCounter,
-      movieAmount: movieCounter,
-    };
-  }),
-
   watchHistory: protectedProcedure
     .input(
       z.object({
@@ -288,96 +204,4 @@ export const profileRouter = router({
         pagesAmount: Math.ceil(sortedHistory.length / input.pageSize),
       };
     }),
-
-  upNext: protectedProcedure.query(async ({ ctx }) => {
-    const episodes = await ctx.prisma.episodesHistory.findMany({
-      where: { user_id: ctx.session.user.profile.userId },
-      include: {
-        series: true,
-      },
-      orderBy: {
-        datetime: "desc",
-      },
-      distinct: ["series_id"],
-    });
-
-    const result = await Promise.all(
-      episodes.flatMap(async (lastEpisode) => {
-        const season = await ctx.prisma.seasons.findFirst({
-          where: {
-            series_id: lastEpisode.series_id,
-            season_number: lastEpisode.season_number,
-          },
-          include: {
-            episodes: true,
-            Series: true,
-          },
-        });
-
-        // Removes all episodes that don't have a next episode
-        const nextEpisode = season?.episodes.filter((ep) => {
-          if (
-            ep.episode_number === lastEpisode.episode_number + 1 &&
-            ep.season_number === lastEpisode.season_number &&
-            ep?.air_date !== null &&
-            ep?.air_date <= new Date()
-          ) {
-            return true;
-          } else {
-            return false;
-          }
-        });
-
-        if (nextEpisode !== undefined && nextEpisode?.length > 0) {
-          return {
-            ...nextEpisode[0],
-            series: season?.Series,
-          };
-        } else {
-          const nextSeason = await ctx.prisma.seasons.findFirst({
-            where: {
-              series_id: lastEpisode.series_id,
-              season_number: lastEpisode.season_number + 1,
-            },
-            include: {
-              episodes: true,
-              Series: true,
-            },
-          });
-
-          const nextEpisode = nextSeason?.episodes.filter((ep) => {
-            if (
-              ep.episode_number === 1 &&
-              ep.season_number === lastEpisode.season_number + 1 &&
-              ep?.air_date !== null &&
-              ep?.air_date <= new Date()
-            ) {
-              return true;
-            } else {
-              return false;
-            }
-          });
-
-          if (nextEpisode !== undefined && nextEpisode?.length > 0) {
-            return {
-              ...nextEpisode[0],
-              series: season?.Series,
-            };
-          }
-        }
-      })
-    );
-
-    return {
-      result: paginate(
-        result.filter((el) => {
-          if (el !== undefined) {
-            return true;
-          }
-        }),
-        6,
-        1
-      ),
-    };
-  }),
 });
